// nodes/Bigcommerce/webdav/types.ts
export enum AuthType {
    Auto = "auto",
    Digest = "digest",
    None = "none",
    Password = "password",
    Token = "token"
}
// ...
interface DAVResultBaseResponse {
    href: string;
    responsedescription?: string;
}

export interface DAVResultPropstatResponse extends DAVResultBaseResponse {
    propstat: DAVPropStat;
}

export interface DAVResultStatusResponse extends DAVResultBaseResponse {
    status: string;
}

export type DAVResultResponse = DAVResultBaseResponse &
    Partial<DAVResultPropstatResponse> &
    Partial<DAVResultStatusResponse>;

export interface DAVResultResponseProps {
    displayname: string;
    resourcetype: {
        collection?: unknown;
    };
    getlastmodified?: string;
    getetag?: string;
    getcontentlength?: string;
    getcontenttype?: string;
    "quota-available-bytes"?: string | number;
    "quota-used-bytes"?: string | number;

    [additionalProp: string]: unknown;
}

export interface DAVResult {
    multistatus: {
        response: Array<DAVResultResponse>;
    };
}

export interface DAVResultRawMultistatus {
    response: DAVResultResponse | [DAVResultResponse];
}

export interface DAVResultRaw {
    multistatus: "" | DAVResultRawMultistatus | [DAVResultRawMultistatus];
}

export interface DigestContext {
    username: string;
    password: string;
    nc: number;
    algorithm: string;
    hasDigestAuth: boolean;
    cnonce?: string;
    nonce?: string;
    realm?: string;
    qop?: string;
    opaque?: string;
    ha1?: string;
}
// ...
export interface FileStat {
    filename: string;
    basename: string;
    lastmod: string;
    size: number;
    type: "file" | "directory";
    etag: string | null;
    mime?: string;
    props?: DAVResultResponseProps;
}

export interface SearchResult {
    truncated: boolean;
    results: FileStat[];
}

export interface DAVCompliance {
    compliance: string[];
    server: string;
}

export interface GetDirectoryContentsOptions extends WebDAVMethodOptions {
    deep?: boolean;
    details?: boolean;
    glob?: string;
    includeSelf?: boolean;
}
// ...
export interface Headers {
    [key: string]: string;
}

export interface LockOptions extends WebDAVMethodOptions {
    refreshToken?: string;
    timeout?: string;
}

export interface LockResponse {
    serverTimeout: string;
    token: string;
}

export interface OAuthToken {
    access_token: string;
    token_type: string;
    refresh_token?: string;
}

export interface ProgressEvent {
    loaded: number;
    total: number;
}

export type ProgressEventCallback = (e: ProgressEvent) => void;

export interface PutFileContentsOptions extends WebDAVMethodOptions {
    contentLength?: boolean | number;
    overwrite?: boolean;
    onUploadProgress?: UploadProgressCallback;
}

export type RequestDataPayload = string | Buffer | ArrayBuffer | Record<string, any>;
// ...
export interface ResponseDataDetailed<T> {
    data: T;
    headers: Headers;
    status: number;
    statusText: string;
}
// ...
export interface WebDAVClientContext {
    authType: AuthType;
    remoteBasePath?: string;
    contactHref: string;
    digest?: DigestContext;
    ha1?: string;
    headers: Headers;
    httpAgent?: any;
    httpsAgent?: any;
    parsing: WebDAVParsingContext;
    password?: string;
    remotePath: string;
    remoteURL: string;
    token?: OAuthToken;
    username?: string;
    withCredentials?: boolean;
}
// ...
export interface WebDAVMethodOptions {
    data?: RequestDataPayload;
    headers?: Headers;
    signal?: AbortSignal;
}

/**
 * Callback to parse a prop attribute value.
 * If `undefined` is returned the original text value will be used.
 * If the unchanged value is returned the default parsing will be applied.
 * Otherwise the returned value will be used.
 */
export type WebDAVAttributeParser = (
    jPath: string,
    attributeValue: string
) => string | unknown | undefined;

/**
 * Callback to parse a prop tag value.
 * If `undefined` is returned the original text value will be used.
 * If the unchanged value is returned the default parsing will be applied.
 * Otherwise the returned value will be used.
 */
export type WebDAVTagParser = (jPath: string, tagValue: string) => string | unknown | undefined;

export interface WebDAVParsingContext {
    attributeNamePrefix?: string;
    attributeParsers: WebDAVAttributeParser[];
    tagParsers: WebDAVTagParser[];
}